Rest of Q4 + Moving class to separate file
Rest of Q4 added and carrying on smoothly, runtime is becoming an issue for the multiple scenarios but this seems to be dependent on the grid size, and hopefully will be dealt with further with parallelisation. Also due to the large amount of code, to attempt to streamline things the code for the new Poisson class is moved to its own code file and imported in.