Monte Carlo Changes
Sadly needed to change the monte carlo due to the output of arrays from the poisson code. A shame due to mark loss but easiest way found to make all code work and doesn't seem to be enough time to find a way without changing monte carlo. Biggest and main changes are the addition of an averaging function and a parallelisation which give arrays as outputs instead of summing them as the original functions do.